ok
Our project aims to try to find these children in a local area and send them back to their
families. We target phones as it’s easy to use them and anyone these days has a smart
phone. We made our app works in a local area and we selected the malls as a begin to our
project, cause we can control the area of the mall instead of a public street where we don’t
have a lot of cameras